+++ src-docs/src/views/combo_box/combo_box_delimiter.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 
 import { EuiComboBox } from '../../../../src/components';
-import { DisplayToggles } from '../form_controls/display_toggles';
 
 export default class extends Component {
   constructor(props) {
@@ -83,18 +82,16 @@ export default class extends Component {
   render() {
     const { selectedOptions } = this.state;
     return (
-      /* DisplayToggles wrapper for Docs only */
-      <DisplayToggles canDisabled={false} canReadOnly={false}>
-        <EuiComboBox
-          placeholder="Select or create options"
-          options={this.options}
-          selectedOptions={selectedOptions}
-          onChange={this.onChange}
-          onCreateOption={this.onCreateOption}
-          isClearable={true}
-          data-test-subj="demoComboBox"
-        />
-      </DisplayToggles>
+      <EuiComboBox
+        placeholder="Select or create options"
+        options={this.options}
+        delimiter=","
+        selectedOptions={selectedOptions}
+        onChange={this.onChange}
+        onCreateOption={this.onCreateOption}
+        isClearable={true}
+        data-test-subj="demoComboBox"
+      />
     );
   }
 }